+++ lib/simple_json_serializer_generator.dart
@@ -17,10 +17,11 @@ class SimpleJsonSerializerGenerator {
           final classData = _getValue<Map>(classes, className);
           for (final key in classData.keys) {
             final fieldName = '$key';
-            final fieldType = _getValue<String>(classData, fieldName);
+            final filedData = _getFieldData(classData, fieldName);
+            final type = _getValue<String>(filedData, 'type');
             b.fields.add(Field((b) {
               b.name = fieldName;
-              b.type = Reference(fieldType);
+              b.type = Reference(type);
               b.modifier = FieldModifier.final$;
             }));
           }
@@ -53,9 +54,9 @@ class SimpleJsonSerializerGenerator {
         final enum_ = Enum((b) {
           final enumName = '$key';
           b.name = enumName;
-          final enumData = _getValue<List>(enums, enumName);
-          for (final value in enumData) {
-            final valueName = '$value';
+          final enumData = _getValue<Map>(enums, enumName);
+          for (final key in enumData.keys) {
+            final valueName = '$key';
             b.values.add(EnumValue((b) {
               b.name = valueName;
             }));
@@ -72,16 +73,12 @@ class SimpleJsonSerializerGenerator {
   }
 
   String generateSerializerCollection(
-      String name, Iterable<String> named, Iterable<String> other) {
+      String name, Iterable<String> serializers) {
     final buffer = StringBuffer();
     buffer.writeln('final $name = JsonSerializerCollection()');
     final list = <String>[];
-    for (final name in named) {
-      list.add('..addSerializer(${name}Serializer())');
-    }
-
-    for (final name in other) {
-      list.add('..addSerializer($name())');
+    for (final serializer in serializers) {
+      list.add('..addSerializer($serializer())');
     }
 
     list.sort();
@@ -91,12 +88,69 @@ class SimpleJsonSerializerGenerator {
     return result;
   }
 
-  String generateSerializersForClasses(Map classes) {
+  String generateSerializers(Map serializers) {
+    final library = Library((lib) {
+      for (final key in serializers.keys) {
+        final class_ = Class((b) {
+          final serializerName = '$key';
+          final serializerData = _getValue<Map>(serializers, key);
+          b.name = serializerName;
+          final type = _getValue<String>(serializerData, 'type');
+          b.extend = Reference('JsonSerializer<$type>');
+          b.methods.add(Method((b) {
+            b.annotations.add(CodeExpression(Code('override')));
+            b.name = 'deserialize';
+            b.returns = Reference(type);
+            b.requiredParameters.add(Parameter((b) {
+              b.name = 'deserializer';
+              b.type = Reference('Deserializer');
+            }));
+            b.requiredParameters.add(Parameter((b) {
+              b.name = 'value';
+              b.type = Reference('Object?');
+            }));
+
+            final deserialize =
+                _getValue<String>(serializerData, 'deserialize');
+            b.body = Code(deserialize);
+          }));
+
+          b.methods.add(Method((b) {
+            b.annotations.add(CodeExpression(Code('override')));
+            b.name = 'serialize';
+            final returns =
+                _getValue<String>(serializerData, 'serializeReturns');
+            b.returns = Reference(returns);
+            b.requiredParameters.add(Parameter((b) {
+              b.name = 'serializer';
+              b.type = Reference('Serializer');
+            }));
+            b.requiredParameters.add(Parameter((b) {
+              b.name = 'value';
+              b.type = Reference(type);
+            }));
+
+            final serialize = _getValue<String>(serializerData, 'serialize');
+            b.body = Code(serialize);
+          }));
+        });
+
+        lib.body.add(class_);
+      }
+    });
+
+    final emitter = DartEmitter();
+    final result = library.accept(emitter).toString();
+    return result;
+  }
+
+  String generateSerializersForClasses(
+      Map classes, String Function(String name) generateName) {
     final library = Library((lib) {
       for (final key in classes.keys) {
         final class_ = Class((b) {
           final className = '$key';
-          b.name = '${className}Serializer';
+          b.name = generateName(className);
           b.extend = Reference('JsonSerializer<$className>');
           final classData = _getValue<Map>(classes, className);
           b.methods.add(Method((b) {
@@ -117,16 +171,20 @@ class SimpleJsonSerializerGenerator {
             code.add('return $className(');
             for (final key in classData.keys) {
               final fieldName = '$key';
-              final fieldType = _getValue<String>(classData, fieldName).trim();
+              final fieldData = _getFieldData(classData, fieldName);
+              final type = _getValue<String>(fieldData, 'type').trim();
+              final alias =
+                  (_tryGetValue<String>(fieldData, 'alias') ?? fieldName)
+                      .trim();
               var collection = '';
-              if (fieldType.startsWith('List<')) {
+              if (type.startsWith('List<')) {
                 collection = 'List';
-              } else if (fieldType.startsWith('Map<')) {
+              } else if (type.startsWith('Map<')) {
                 collection = 'Map';
               }
 
               code.add(
-                  "$fieldName: deserializer.deserialize$collection(json['$fieldName']),");
+                  "$fieldName: deserializer.deserialize$collection(json['$alias']),");
             }
 
             code.add(');');
@@ -150,16 +208,20 @@ class SimpleJsonSerializerGenerator {
             code.add('return {');
             for (final key in classData.keys) {
               final fieldName = '$key';
-              final fieldType = _getValue<String>(classData, fieldName).trim();
+              final fieldData = _getFieldData(classData, fieldName);
+              final type = _getValue<String>(fieldData, 'type').trim();
+              final alias =
+                  (_tryGetValue<String>(fieldData, 'alias') ?? fieldName)
+                      .trim();
               var collection = '';
-              if (fieldType.startsWith('List<')) {
+              if (type.startsWith('List<')) {
                 collection = 'List';
-              } else if (fieldType.startsWith('Map<')) {
+              } else if (type.startsWith('Map<')) {
                 collection = 'Map';
               }
 
               code.add(
-                  "'$fieldName': serializer.serialize$collection(value.$fieldName),");
+                  "'$alias': serializer.serialize$collection(value.$fieldName),");
             }
 
             code.add('};');
@@ -176,13 +238,23 @@ class SimpleJsonSerializerGenerator {
     return result;
   }
 
-  String generateSerializersForEnums(Map enums) {
+  String generateSerializersForEnums(
+      Map enums, String Function(String name) generateName) {
     final library = Library((lib) {
       for (final key in enums.keys) {
         final class_ = Class((b) {
           final enumName = '$key';
-          b.name = '${enumName}Serializer';
+          b.name = generateName(enumName);
           b.extend = Reference('JsonSerializer<$enumName>');
+          final enumData = _getValue<Map>(enums, enumName);
+          final values = <String, dynamic>{};
+          for (final key in enumData.keys) {
+            final valueName = '$key'.trim();
+            final value = _tryGetValue(enumData, valueName);
+            values[valueName] = value;
+          }
+
+          final isMapped = values.values.any((e) => e != null);
           b.methods.add(Method((b) {
             b.annotations.add(CodeExpression(Code('override')));
             b.name = 'deserialize';
@@ -197,15 +269,34 @@ class SimpleJsonSerializerGenerator {
             }));
 
             final code = <String>[];
-            code.add('final json = cast<int>(value);');
-            code.add('return $enumName.values[json];');
+            if (!isMapped) {
+              code.add('final json = cast<int>(value);');
+              code.add('return $enumName.values[json];');
+            } else {
+              final entries = <String>[];
+              var index = 0;
+              for (final key in values.keys) {
+                final value = values[key];
+                final val = value ?? index;
+                final entry = '$val: $index';
+                entries.add(entry);
+                index++;
+              }
+
+              code.add('const map = {');
+              code.add(entries.join(',\n'));
+              code.add('};');
+              code.add('final index = map[value] as int;');
+              code.add('return $enumName.values[index];');
+            }
+
             b.body = Code(code.join('\n'));
           }));
 
           b.methods.add(Method((b) {
             b.annotations.add(CodeExpression(Code('override')));
             b.name = 'serialize';
-            b.returns = Reference('int');
+            b.returns = isMapped ? Reference('Object?') : Reference('int');
             b.requiredParameters.add(Parameter((b) {
               b.name = 'serializer';
               b.type = Reference('Serializer');
@@ -216,7 +307,25 @@ class SimpleJsonSerializerGenerator {
             }));
 
             final code = <String>[];
-            code.add('return value.index;');
+            if (!isMapped) {
+              code.add('return value.index;');
+            } else {
+              final entries = <String>[];
+              var index = 0;
+              for (final key in values.keys) {
+                final value = values[key];
+                final val = value ?? index;
+                final entry = '$index: $val';
+                entries.add(entry);
+                index++;
+              }
+
+              code.add('const map = {');
+              code.add(entries.join(',\n'));
+              code.add('};');
+              code.add('return map[value.index];');
+            }
+
             b.body = Code(code.join('\n'));
           }));
         });
@@ -230,6 +339,16 @@ class SimpleJsonSerializerGenerator {
     return result;
   }
 
+  Map _getFieldData(Map map, String key) {
+    final type = _tryGetValue<String>(map, key);
+    if (type != null) {
+      return {'type': type};
+    }
+
+    final result = _getValue<Map>(map, key);
+    return result;
+  }
+
   T _getValue<T>(Map map, Object? key) {
     final result = map[key];
     if (result is T) {
@@ -238,5 +357,14 @@ class SimpleJsonSerializerGenerator {
 
     throw StateError(
         "Unable to cast field ${result.runtimeType} '$key' to type $T");
+  }
+
+  T? _tryGetValue<T>(Map map, Object? key) {
+    final result = map[key];
+    if (result is T) {
+      return result;
+    }
+
+    return null;
   }
 }